# backend/src/routes/folders.ts
// Ordner-Endpunkte — neu durch die Contract-Änderung "benutzerdefinierte
// Ordner statt festem Enum" (SYNC.md, Commit 734781e). Siehe api-spec.yaml
// /folders, /folders/{folderId}.

import { Router } from "express";
import { store } from "../db/store";
import { toApiFolder } from "../mappers";
import type { SystemFolderKey } from "../types";

export const foldersRouter = Router();

// quarantaene/spam/papierkorb/entwuerfe/gesendet sind laut design-tokens.json
// (systemFolders.defaults, renamable: false) nicht umbenennbar — API lehnt
// PATCH auf `name` für diese mit 400 ab. [2026-09-10] Ordner-Umbau
// (WEB_INBOX.md 09.09.): entwuerfe/gesendet ergänzt.
const NOT_RENAMABLE: SystemFolderKey[] = ["quarantaene", "spam", "papierkorb", "entwuerfe", "gesendet"];

// design-tokens.json: customFolder.defaultIcon
const CUSTOM_FOLDER_DEFAULT_ICON = "folder";

// GET /folders — siehe api-spec.yaml
foldersRouter.get("/folders", async (req, res) => {
  res.json((await store.listFolders(req.userId)).map(toApiFolder));
});

// POST /folders — eigenen Ordner anlegen (is_system=false, system_key=null)
foldersRouter.post("/folders", async (req, res) => {
  const name = typeof req.body?.name === "string" ? req.body.name.trim() : "";
  if (!name) return res.status(400).json({ error: "name ist erforderlich" });

  const icon =
    typeof req.body?.icon === "string" && req.body.icon.trim() ? req.body.icon.trim() : CUSTOM_FOLDER_DEFAULT_ICON;

  // sort_order: ans Ende der bestehenden Liste anhängen (kein sortOrder im
  // Request-Body laut api-spec.yaml POST /folders — nur PATCH erlaubt das).
  const sortOrder = (await store.listFolders(req.userId)).length;

  const folder = await store.createFolder({
    userId: req.userId,
    name,
    icon,
    isSystem: false,
    systemKey: null,
    sortOrder,
  });
  res.status(201).json(toApiFolder(folder));
});

// PATCH /folders/:folderId — umbenennen/Icon/Reihenfolge ändern
foldersRouter.patch("/folders/:folderId", async (req, res) => {
  const folder = await store.getFolder(req.params.folderId);
  if (!folder) return res.status(404).json({ error: "Ordner nicht gefunden" });
  // [2026-09-10] echte Auth: Besitz-Prüfung, vorher fehlte die komplett
  // (gab bis dahin ohnehin nur den einen Demo-User, an dessen Daten nur er
  // selbst überhaupt einen Request stellen konnte).
  if (folder.userId !== req.userId) return res.status(403).json({ error: "Ordner gehört nicht zum angemeldeten User" });

  const patch: { name?: string; icon?: string; sortOrder?: number } = {};

  if (req.body?.name !== undefined) {
    if (folder.isSystem && folder.systemKey && NOT_RENAMABLE.includes(folder.systemKey)) {
      return res.status(400).json({ error: `Systemordner '${folder.systemKey}' kann nicht umbenannt werden` });
    }
    if (typeof req.body.name !== "string" || !req.body.name.trim()) {
      return res.status(400).json({ error: "name darf nicht leer sein" });
    }
    patch.name = req.body.name.trim();
  }

  if (req.body?.icon !== undefined) {
    if (typeof req.body.icon !== "string" || !req.body.icon.trim()) {
      return res.status(400).json({ error: "icon darf nicht leer sein" });
    }
    patch.icon = req.body.icon.trim();
  }

  if (req.body?.sortOrder !== undefined) {
    if (typeof req.body.sortOrder !== "number") {
      return res.status(400).json({ error: "sortOrder muss eine Zahl sein" });
    }
    patch.sortOrder = req.body.sortOrder;
  }

  const updated = (await store.updateFolder(folder.id, patch))!;
  res.json(toApiFolder(updated));
});

// DELETE /folders/:folderId — eigenen Ordner löschen (System-Ordner nicht löschbar)
foldersRouter.delete("/folders/:folderId", async (req, res) => {
  const folder = await store.getFolder(req.params.folderId);
  if (!folder) return res.status(404).json({ error: "Ordner nicht gefunden" });
  if (folder.userId !== req.userId) return res.status(403).json({ error: "Ordner gehört nicht zum angemeldeten User" });

  if (folder.isSystem) {
    return res.status(400).json({ error: "Systemordner können nicht gelöscht werden" });
  }

  // Design-Entscheidung (nicht im Contract festgelegt, siehe SYNC.md/README):
  // messages.folder_id ist im Schema NOT NULL/FK, darf also nie ins Leere
  // zeigen. Nachrichten aus dem gelöschten Ordner werden deshalb vor dem
  // Löschen in den System-Ordner "sonstiges" verschoben.
  const fallback = await store.getSystemFolder(folder.userId, "sonstiges");
  if (fallback) {
    for (const message of await store.listMessages({ folderId: folder.id })) {
      await store.moveMessage(message.id, fallback.id);
    }
  }

  await store.deleteFolder(folder.id);
  res.status(204).send();
});