# contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: driftmail API
  version: 0.1.0
  description: >
    Track 0 Contract. Track A implementiert diese Spec real.
    Track C und F bauen zunächst gegen einen Mock-Server, der exakt
    diese Response-Formate zurückgibt.
servers:
  - url: https://api.driftware.online/v1
security:
  - bearerAuth: []
paths:
  /auth/session:
    post:
      summary: >
        Session-Token erneuern. Login selbst passiert implizit durch den
        OAuth-Flow beim Verbinden eines Mail-Kontos (siehe POST /accounts),
        dieser Endpoint dient nur der Token-Erneuerung. `security: []`
        bedeutet nur, dass ein bereits abgelaufener Token hier ausnahmsweise
        nicht per globalem `bearerAuth`-Mechanismus mit 401 abgewiesen wird,
        bevor der Handler ihn sieht -- der (noch nicht abgelaufene, bzw.
        innerhalb der serverseitigen Kulanzfrist liegende) aktuelle Token
        muss trotzdem im `Authorization`-Header mitgeschickt werden, sonst
        401. [2026-09-10] Terminal, siehe backend/README.md "Auth": bisher
        war dieser Endpoint zwar spezifiziert, aber von keinem Code
        implementiert (wie zuvor schon bei /messages/{messageId}/unsubscribe).
      security: []
      responses:
        "200":
          description: Neues Token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
        "401":
          description: Kein/ungueltiger/abgelaufener Token im Authorization-Header
  /accounts:
    get:
      summary: Verbundene Mail-Konten des angemeldeten Users auflisten
      responses:
        "200":
          description: Liste der Mail-Konten
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MailAccount"
    post:
      summary: >
        Mail-Konto verbinden -- das ist zugleich der Login/Registrierungs-
        Weg (siehe /auth/session-Beschreibung). [2026-09-10] Terminal,
        Contract-Luecke geschlossen: dieser Endpoint fehlte komplett, obwohl
        /auth/session bereits seit Track 0 darauf verwies. Legt bei erster
        Verbindung dieser E-Mail-Adresse einen neuen User + dessen 7
        Standard-Ordner an, sonst wird der bestehende User/das bestehende
        Konto wiederverwendet (kein Multi-Account pro User in diesem
        Entwicklungsstand, siehe backend/README.md "Auth"). Echter
        OAuth-Code-Austausch (Gmail) bzw. echte IMAP-Zugangsdaten-Pruefung
        sind noch nicht angebunden (`oauthCode`/`imapPassword` werden
        aktuell nicht ausgewertet) -- bewusste, dokumentierte Grenze, analog
        zum bestehenden Fixture-Adapter-Muster fuer den Mail-Sync selbst.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [provider, emailAddress]
              properties:
                provider: { type: string, enum: [gmail, imap] }
                emailAddress: { type: string }
                oauthCode: { type: string, description: "Gmail: noch nicht ausgewertet, siehe summary" }
                imapPassword: { type: string, description: "IMAP: noch nicht ausgewertet, siehe summary" }
      responses:
        "200":
          description: Konto verbunden, neuer Session-Token ausgestellt
          content:
            application/json:
              schema:
                type: object
                properties:
                  account:
                    $ref: "#/components/schemas/MailAccount"
                  token: { type: string }
        "400":
          description: provider oder emailAddress fehlt/ungueltig
  /folders:
    get:
      summary: Ordner des Users auflisten (System- und eigene Ordner)
      responses:
        "200":
          description: Liste der Ordner
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Folder"
    post:
      summary: Neuen eigenen Ordner anlegen
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                icon: { type: string }
      responses:
        "201":
          description: Angelegt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
  /folders/{folderId}:
    patch:
      summary: Ordner umbenennen/Icon ändern (System-Ordner außer quarantaene/spam ebenfalls umbenennbar)
      parameters:
        - name: folderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                icon: { type: string }
                sortOrder: { type: integer }
      responses:
        "200":
          description: Aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
    delete:
      summary: Eigenen Ordner löschen (System-Ordner nicht löschbar)
      parameters:
        - name: folderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Gelöscht
...
  /drafts:
    get:
      summary: Entwuerfe des Users auflisten (Inhalt des "entwuerfe"-Systemordners)
      responses:
        "200":
          description: Liste der Entwuerfe
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Draft"
    post:
      summary: >
        Neuen (leeren oder vorbefuellten) Entwurf anlegen -- beim Start
        eines Compose-Vorgangs (neue Mail oder Antwort). Siehe WEB_INBOX.md
        09.09. "KORREKTUR/ERWEITERUNG des Ordner-Umbau-Eintrags".
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                inReplyToMessageId: { type: string, format: uuid, nullable: true }
                to:
                  type: array
                  items: { type: string }
                cc:
                  type: array
                  items: { type: string }
                subject: { type: string }
                bodyText: { type: string }
      responses:
        "200":
          description: Angelegter Entwurf
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
  /drafts/{draftId}:
    parameters:
      - name: draftId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      summary: Laufendes Speichern waehrend des Tippens oder beim Verlassen des Compose-Screens
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                to:
                  type: array
                  items: { type: string }
                cc:
                  type: array
                  items: { type: string }
                subject: { type: string }
                bodyText: { type: string }
      responses:
        "200":
          description: Aktualisierter Entwurf
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
        "404":
          description: Entwurf nicht gefunden
    delete:
      summary: Entwurf verwerfen (Auto-Loeschung nach erfolgreichem Senden, siehe POST /messages/send draftId)
      responses:
        "200":
          description: Geloescht
        "404":
          description: Entwurf nicht gefunden
...
  /contracts/{contractId}/confirm:
    post:
      summary: Erkannte Vertragsdaten bestätigen oder korrigieren
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Contract"
      responses:
        "200":
          description: Bestätigt
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Session-Token nach OAuth-Verknuepfung des Mail-Kontos. userId wird
        serverseitig aus dem Token aufgeloest, nie aus Request-Body/Query/Pfad
        vertraut.
  schemas:
    Folder:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        icon: { type: string }
        isSystem: { type: boolean }
        systemKey:
          type: string
          nullable: true
          enum: [eingang, entwuerfe, gesendet, sonstiges, quarantaene, spam, papierkorb]
        sortOrder: { type: integer }
    Draft:
      type: object
      properties:
        id: { type: string, format: uuid }
        inReplyToMessageId: { type: string, format: uuid, nullable: true }
        to:
          type: array
          items: { type: string }
        cc:
          type: array
          items: { type: string }
        subject: { type: string, nullable: true }
        bodyText: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time }
    MailAccount:
      type: object
      properties:
        id: { type: string, format: uuid }
        provider: { type: string, enum: [gmail, imap] }
        emailAddress: { type: string }
        syncStatus: { type: string, enum: [pending, syncing, ok, error] }
...
    Contract:
      type: object
      properties:
        id: { type: string, format: uuid }
        providerName: { type: string }
        contractStart: { type: string, format: date, nullable: true }
        contractEnd: { type: string, format: date, nullable: true }
        cancellationDeadline: { type: string, format: date, nullable: true }
        cancellationPeriodDays: { type: integer, nullable: true }
        status: { type: string, enum: [active, cancelled, expired, needs_review] }
        extractedConfidence: { type: number, format: float }
...